Improve loinc code choice for lab result id
The previous code was for a case identifier and used multiple times.
This new loinc code is specific to lab results.
Another option could be 19146-0 which relates to reference lab test
results - I did not use this one however as it does not specify that it
is an identifier field - which 30896-5 does.

https://loinc.org/30896-5/
https://loinc.org/19146-0/
https://jembiprojects.jira.com/browse/DSC19-84