Improve overall test result loinc code in lab result
The previous loinc code referenced results of a very specific nucleic
acid assay that could confirm presence of covid.
This is too specific. We do not know what test method was used for
presence confirmation - since the accuracy of different test methods is
so variable it is important that the loinc code not give an innaccurate
reading of accuracy.

https://loinc.org/19146-0/
https://jembiprojects.jira.com/browse/DSC19-84