Replace dateTime observation with field in DiagnosticReport
The purpose of the FHIR transaction Bundle containing the resources is
to give clinical context of an event. In the lab report event the result
of the report was listed. Previously this value was displayed as an
observation that was not referenced in any of the other resources in the
transation Bundle - this piece of information would be missing its
clinical relevance if the entire Bundle wasn't read (since it is a
transaction Bundle it wouldn\t be read in this manner - this would need
to be a document Bundle to function in that way)
Therefore, to ensure the lab report date context was kept, the date will
be added to the diagnortic report effectiveDateTime. This field aims to
document the "time of the procedure" in this case the reporting.

https://jembiprojects.jira.com/browse/DSC19-85