Modify the profiles
The valueset that was being used for the countries was incorrect and has been corrected. Some observation profiles have been changed to use the valuestring, valueDatetime or valueCodeablecConcept, and not the component field. This is for consistency as all other observation profiles do not use the component which would result in us having to add the observation code twice.

DSC19-71